Refactor to eliminate circular dependency problems
